requirements.txt
numpy~=1.26.4
pandas~=2.2.1
streamlit~=1.31.1
Pillow~=10.2.0
matplotlib~=3.8.3
scipy~=1.12.0
scikit-learn~=1.4.1.post1
joblib~=1.3.2
imblearn
langdetect~=1.0.9
keras~=3.3.3
nltk~=3.8.1
xgboost~=2.0.3
protobuf~=3.20.0
xgboost~=2.0.3
swifter~=1.4.0
spacy~=3.7.3
seaborn~=0.13.2
swifter
wordcloud~=1.9.3